facial_recognition/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py38h32f6830_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=py_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h1e0a361_1000
c-ares=1.16.1=h516909a_3
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.2=py38he30daa8_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=pyh9f0ad1d_0
cryptography=3.1=py38h766eaa4_0
cycler=0.10.0=py_2
flask=1.1.2=pypi_0
freetype=2.10.2=h5ab3b9f_0
gast=0.3.3=py_0
google-auth=1.21.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.31.0=py38h2c89da0_0
h5py=2.10.0=nompi_py38hfb01d0b_104
hdf5=1.10.6=nompi_h3c11f04_101
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py38h32f6830_0
intel-openmp=2020.2=254
itsdangerous=1.1.0=pypi_0
jinja2=2.11.2=pypi_0
jpeg=9b=habf39ab_1
keras=2.4.3=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.2.0=py38hbf85e49_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.1=1
matplotlib-base=3.3.1=py38h91b0d89_1
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
opencv-python=4.4.0.42=pypi_0
openssl=1.1.1g=h516909a_1
opt_einsum=3.3.0=py_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.2=py38_0
protobuf=3.13.0=py38h950e882_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pyyaml=5.3.1=py38h1e0a361_0
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scipy=1.5.2=py38h0b6359f_0
setuptools=49.6.0=py38_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tensorflow=2.2.0=mkl_py38h6d3daf0_0
tensorflow-base=2.2.0=mkl_py38h5059a2d_0
tensorflow-estimator=2.2.0=pyh95af2a2_0
termcolor=1.1.0=py_2
tk=8.6.10=hbc83047_0
tornado=6.0.4=py38h1e0a361_1
urllib3=1.25.10=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=py_0
wrapt=1.12.1=py38h1e0a361_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
